Config: pull config helpers out into module
This way not all of `Configuration` is included in classes.
