+++ lib/baes.rb
@@ -6,6 +6,7 @@ module Baes; end
 class Baes::Error < StandardError; end
 
 require_relative "baes/configuration"
+require_relative "baes/configuration/helpers"
 
 require_relative "baes/actions"
 require_relative "baes/branch"
