+++ lib/baes/configuration.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "optparse"
-
 # module to allow configuring various Baes settings
 module Baes::Configuration
   # return the git wrapper in use, Baes::Git by default
@@ -72,36 +70,6 @@ module Baes::Configuration
   # allow setting auto skip mode
   def self.auto_skip=(auto_skip)
     @auto_skip = auto_skip
-  end
-
-  # return the configured git wrapper
-  def git
-    Baes::Configuration.git
-  end
-
-  # return the configured input
-  def input
-    Baes::Configuration.input
-  end
-
-  # return the configured output
-  def output
-    Baes::Configuration.output
-  end
-
-  # return the configured root name if given
-  def root_name
-    Baes::Configuration.root_name
-  end
-
-  # return the configured ignored branch names
-  def ignored_branch_names
-    Baes::Configuration.ignored_branch_names
-  end
-
-  # return whether dry run has been enabled
-  def dry_run?
-    Baes::Configuration.dry_run?
   end
 
   # clear all configuration
